refactor controllable to player
